EN-19215: Skip grouping when querying the rollup if it already matches the query grouping.
In order to allow using rollups to speed up common grouped queries, we want to be able to skip
doing the grouping if the rollup is already grouped the exact way the query wants.  This doesn't
matter for the smaller rollups that the code was originally designed to support but for other
uses that are large after grouping it makes a big difference.