EN-19741: Add rewrite support for HAVING.
The base case for HAVING is pretty simple, as expressions there are something
we already are mapping in the selection or grouping so they map the same way.

In the case where the grouping in the query has the exact same columns as the rollup
and we are removing the grouping, we have to move the condition into the where clause.