EN-40314: Don't rewrite window functions.
Window functions run after where / group by / having so can't be
rewritten in isolation.  We could still rewrite a query with a
window function in if the other clauses all match, however that
requires coordination between expression mapping and mapping other
clauses at a higher level that isn't implemented, so for now we
just forbid it entirely to avoid incorrect rewrites.