EN-51750: Support join queries in rollups (#251)
This is a general effort in improving query performance by supporting a wider variety of queries in rollups.
Join queries can use rollups when their joins match.
Rollup updated-ness is out of scope.  It requires request header X-Socrata-Debug: true to use rollups with joins
This is to prevent people burnt by outdated rollups because we do not refresh a rollup when join tables are updated.