NOJIRA: Separating QueryRewriter interface and implementation (#278)
QueryRewriter is a large class and it is a dependency used by QueryResource. QueryResource only uses (and so needs to now) a few methods from QueryRewriter - i.e. an interface. It is better to separate an interface from the implementation - less coupling, clear interface definition, easier to understand code.

With readability and code organization in mind, classes related to large functional block (i.e. rollups) are better placed in a package.

The changes made in this PR:

QueryRewriter becomes an abstract class defining the interface
QueryRewriterImplementation implements it
files related to the rollups moved to query coordinator.rollups
some utility methods moved from companion to class

Reviewed-by: Robert Voyer @rlvoyer