Retry instead of failing when not found in the chosen secondary
This can happen if the dataset is moved out from under us (e.g., because
of a collocation).  This can turn a "not found" reponse into either a
"ran out of retries" or a "dataset not available" if the dataset is
genuinely gone - but if the dataset had been gone and not in our cache
in the first place, it would've been "dataset not available" anyway.